Distribution Patterns
Iris-setosa: The sepal lengths of Iris-setosa primarily cluster around 5 to 6 cm. The highest frequency of sepal lengths is around 5.0 to 5.5 cm.

Iris-versicolor: The sepal lengths of Iris-versicolor are more spread out compared to Iris-setosa, with a noticeable peak around 6 cm.

Iris-virginica: The sepal lengths of Iris-virginica show a wider range, generally longer than those of Iris-setosa and Iris-versicolor, with a peak frequency around 6.5 to 7 cm.